+++ src/cmd/pr/approve.rs
@@ -20,7 +20,10 @@ pub fn approve(args: Cli, conf: Configuration) -> Result<()> {
     } = args;
     if let Commands::Pr(PrCommands::Approve {}) = command {
         let repository = LocalRepository::init(dir)?;
-        let (hostname, repo, branch) = repository.get_parsed_remote(branch)?;
+        let (hostname, repo, remote_branch) = repository.get_parsed_remote(branch)?;
+        let remote_branch = remote_branch.wrap_err(eyre!(
+            "You have to push this branch first before you can approve."
+        ))?;
 
         // Find settings or use the auth command
         let settings = conf.find_settings(&hostname, &repo);
@@ -38,7 +41,7 @@ pub fn approve(args: Cli, conf: Configuration) -> Result<()> {
         };
 
         let vcs = init_vcs(hostname, repo, settings)?;
-        let pr = vcs.get_pr_by_branch(&branch)?;
+        let pr = vcs.get_pr_by_branch(&remote_branch)?;
         vcs.approve_pr(pr.id)?;
         pr.print(false, output.into());
         Ok(())