+++ src/cmd/pr/close.rs
@@ -19,7 +19,10 @@ pub fn close(args: Cli, conf: Configuration) -> Result<()> {
     } = args;
     if let Commands::Pr(PrCommands::Close {}) = command {
         let repository = LocalRepository::init(dir)?;
-        let (hostname, repo, branch) = repository.get_parsed_remote(branch)?;
+        let (hostname, repo, remote_branch) = repository.get_parsed_remote(branch)?;
+        let remote_branch = remote_branch.wrap_err(eyre!(
+            "You have to push this branch first before you can close a PR."
+        ))?;
 
         // Find settings or use the auth command
         let settings = conf.find_settings(&hostname, &repo);
@@ -37,7 +40,7 @@ pub fn close(args: Cli, conf: Configuration) -> Result<()> {
         };
 
         let vcs = init_vcs(hostname, repo, settings)?;
-        let pr = vcs.get_pr_by_branch(&branch)?;
+        let pr = vcs.get_pr_by_branch(&remote_branch)?;
         let pr = vcs.close_pr(pr.id)?;
         pr.print(false, output.into());
         Ok(())