+++ src/cmd/pr/create.rs
@@ -3,15 +3,15 @@ use std::io::{stdin, BufRead, Error};
 use crate::cmd::{
     args::{Cli, Commands, OutputType, PrCommands},
     config::{Configuration, RepositoryConfig},
+    pr::merge::merge,
 };
-use colored::Colorize;
 use eyre::{eyre, ContextCompat, Result};
 use gr_bin::vcs::common::{init_vcs, CreatePullRequest};
 use gr_bin::{git::git::LocalRepository, vcs::common::VersionControlSettings};
 use tracing::{debug, info, instrument, trace};
 
 #[instrument(skip_all, fields(command = ?args.command))]
-pub fn create(args: Cli, mut conf: Configuration) -> Result<()> {
+pub fn create(mut args: Cli, mut conf: Configuration) -> Result<()> {
     let Cli {
         command,
         branch,
@@ -19,7 +19,7 @@ pub fn create(args: Cli, mut conf: Configuration) -> Result<()> {
         auth,
         output,
         ..
-    } = args;
+    } = args.clone();
     if let Commands::Pr(PrCommands::Create {
         message,
         description,
@@ -27,11 +27,36 @@ pub fn create(args: Cli, mut conf: Configuration) -> Result<()> {
         delete,
         open,
         reviewers,
-        merge,
+        should_merge,
+        force_merge,
     }) = command
     {
         let repository = LocalRepository::init(dir)?;
-        let (hostname, repo, branch_name) = repository.get_parsed_remote(branch.clone())?;
+        let (hostname, repo, remote_branch) = repository.get_parsed_remote(branch.clone())?;
+
+        // If branch has no remote, we can push it first
+        let remote_branch = remote_branch
+            .wrap_err("You have to push this branch first before you can create a PR.")
+            .or_else(|_| {
+                let remotes = repository.get_remotes()?;
+                let remote = remotes.first().wrap_err(
+                    "There are no remotes, push to a repository before you can create a PR.",
+                )?;
+
+                let branch = if let Some(branch) = &branch {
+                    branch.to_string()
+                } else {
+                    repository.get_branch()?
+                };
+                let message = format!("Branch {branch} is not pushed, pushing to {remote}.");
+                match output {
+                    OutputType::Json => info!("{}", message),
+                    _ => println!("{}", message),
+                };
+                repository.push(remote, &branch)?;
+
+                Ok(branch) as Result<String>
+            })?;
 
         // Find settings or use the auth command
         let settings = conf.find_settings(&hostname, &repo);
@@ -68,7 +93,7 @@ pub fn create(args: Cli, mut conf: Configuration) -> Result<()> {
             .or_else(|| {
                 let commits = repository
                     .get_branch_commits_from_target(
-                        branch,
+                        branch.clone(),
                         target
                             .clone()
                             .or(default_branch)
@@ -99,10 +124,10 @@ pub fn create(args: Cli, mut conf: Configuration) -> Result<()> {
             .unwrap_or_default();
         let is_default_branch = target.is_none();
 
-        let mut pr = vcs.create_pr(CreatePullRequest {
+        let pr = vcs.create_pr(CreatePullRequest {
             title: message,
             description,
-            source: branch_name,
+            source: remote_branch,
             target,
             close_source_branch: delete,
             reviewers: reviewers.unwrap_or_default(),
@@ -111,33 +136,13 @@ pub fn create(args: Cli, mut conf: Configuration) -> Result<()> {
         pr.print(open, output.into());
 
         // Merge the PR instantly if merge is passed
-        if merge {
+        if should_merge {
             info!("Merging pull request {} instantly.", pr.id);
-            pr = vcs.merge_pr(pr.id, false)?;
-
-            let target_branch = pr.target.clone();
-
-            let message = format!(
-                "Checking out to {} and pulling after merge.",
-                target_branch.blue()
-            );
-            match output {
-                OutputType::Json => info!("{}", message),
-                _ => println!("{}", message),
-            };
-            repository.checkout_remote_branch(target_branch, output != OutputType::Json)?;
-
-            // Delete local branch if delete was passed
-            if delete {
-                let source_branch = pr.source;
-                repository.delete_branch(source_branch.clone())?;
-
-                let message = format!("Deleted branch {}.", source_branch.blue());
-                match output {
-                    OutputType::Json => info!("{}", message),
-                    _ => println!("{}", message),
-                };
-            }
+            args.command = Commands::Pr(PrCommands::Merge {
+                delete,
+                force: force_merge,
+            });
+            merge(args, conf.clone())?;
         }
 
         // Save default branch to config for caching