+++ src/cmd/pr/get.rs
@@ -19,7 +19,10 @@ pub fn get(args: Cli, conf: Configuration) -> Result<()> {
     } = args;
     if let Commands::Pr(PrCommands::Get { open }) = command {
         let repository = LocalRepository::init(dir)?;
-        let (hostname, repo, branch) = repository.get_parsed_remote(branch)?;
+        let (hostname, repo, remote_branch) = repository.get_parsed_remote(branch)?;
+        let remote_branch = remote_branch.wrap_err(eyre!(
+            "You have to push this branch first before you can get the PR."
+        ))?;
 
         // Find settings or use the auth command
         let settings = conf.find_settings(&hostname, &repo);
@@ -38,7 +41,7 @@ pub fn get(args: Cli, conf: Configuration) -> Result<()> {
 
         let vcs = init_vcs(hostname, repo, settings)?;
 
-        let pr = vcs.get_pr_by_branch(&branch)?;
+        let pr = vcs.get_pr_by_branch(&remote_branch)?;
         pr.print(open, output.into());
         Ok(())
     } else {