+++ src/vcs/common.rs
@@ -29,6 +29,8 @@ pub struct PullRequest {
     pub description: String,
     pub source: String,
     pub target: String,
+    pub source_sha: String,
+    pub target_sha: String,
     pub url: String,
     #[serde(with = "time::serde::iso8601")]
     pub created_at: OffsetDateTime,
@@ -37,6 +39,7 @@ pub struct PullRequest {
     pub author: User,
     pub closed_by: Option<User>,
     pub reviewers: Option<Vec<User>>,
+    pub delete_source_branch: bool,
 }
 
 impl PullRequest {